Add item variants api & Finish update & destroy item routes & refactor ingredients repository & TODO finish refactor exceptions handling

+++ app/Services/V1/Ingredient/IngredientService.php
@@ -6,8 +6,12 @@ use App\Contracts\V1\Ingredient\IngredientRepositoryInterface;
 use App\Contracts\V1\Ingredient\IngredientServiceInterface;
 use App\DTO\V1\Ingredient\CreateIngredientDTO;
 use App\DTO\V1\Ingredient\UpdateIngredientDTO;
+use App\Exceptions\V1\Ingredient\IngredientCreationException;
+use App\Exceptions\V1\Ingredient\IngredientDeletionException;
+use App\Exceptions\V1\Ingredient\IngredientUpdateException;
 use App\Models\V1\Ingredient;
 use Illuminate\Pagination\LengthAwarePaginator;
+use Throwable;
 
 readonly class IngredientService implements IngredientServiceInterface
 {
@@ -23,18 +27,65 @@ readonly class IngredientService implements IngredientServiceInterface
 
     public function update(Ingredient $ingredient, UpdateIngredientDTO $data): Ingredient
     {
-        return $this->ingredientRepository->update($ingredient, $data);
+        try {
+            $name = $data->name ?? null;
+            if($name && $name !== $ingredient->name && $this->ingredientRepository->findBy('name', $name)){
+                throw IngredientUpdateException::nameAlreadyExists($name);
+            }
+
+            if(isset($data->is_allergen) && $data->is_allergen !== $ingredient->is_allergen && $ingredient->items()->where('is_active', true)->exists()){
+                throw IngredientUpdateException::cannotChangeAllergenStatus();
+            }
+
+            /* @var Ingredient */
+            return $this->ingredientRepository->update($ingredient, collect($data->toArray())->filter(fn($val) => $val !== null)->toArray());
+        } catch (Throwable $e){
+            if($e instanceof IngredientUpdateException){
+                throw $e;
+            }
+
+            throw IngredientUpdateException::default($e->getMessage());
+        }
     }
 
     public function create(CreateIngredientDTO $data): Ingredient
     {
-        return $this->ingredientRepository->createIngredient($data);
+        try {
+            if($this->ingredientRepository->findBy('name', $data->name)){
+                throw IngredientCreationException::nameAlreadyExists($data->name);
+            }
+
+            /* @var Ingredient */
+            return $this->ingredientRepository->create($data->toArray());
+        } catch (Throwable $e){
+            if($e instanceof IngredientCreationException){
+                throw $e;
+            }
+
+            throw IngredientCreationException::default($e->getMessage());
+        }
     }
 
     public function destroy(Ingredient $ingredient): bool
     {
-        $ingredient->items()->detach();
+        try {
+            if ($ingredient->hasActiveItems()){
+                throw IngredientDeletionException::usedInActiveItems();
+            }
 
-        return $this->ingredientRepository->delete($ingredient);
+            if($ingredient->is_mandatory){
+                throw IngredientDeletionException::isMandatoryAllergen();
+            }
+
+            $ingredient->items()->detach();
+
+            return $this->ingredientRepository->delete($ingredient);
+        } catch (Throwable $e){
+            if($e instanceof IngredientDeletionException){
+                throw $e;
+            }
+
+            throw IngredientDeletionException::default($e->getMessage());
+        }
     }
 }
